Create MockPlugin on top of MockConnector
* Remove Tpch usages
* Add missing handles
* Extract inner classes
